test(structs): verify structs marshal and unmarshal from JSON
This is done using Go's golden file convention. Golden files are
generated from the tests when running `go test` with
`-update-golden`. Without that flag, the golden files are read in and
compared against the structs.

This allows for verifying structs can be used to both parse and generate
JSON correctly as it did before code changes, but also allows for easily
updating the JSON blobs when struct schemas change.

And most importantly, this frees up other tests to focus on just making
the right response, and not also verifying that a given response is
parsed correctly.