test(requests): simplify request test data and fixtures
Due now having JSON marshaling tests for every struct used to represent
data from the Katapult API, there's no need to include massive amounts
of data in the request tests as a means of validating the request method
can properly parse the full real-world response. As we already know it
can parse the full set of fields for each type.

Basically, minimal request test data to mostly ensure that we get the
correct object type back with a couple of basic fields for good measure.