feat(client): add queryable interface allowing better object lookup
Many objects can be looked up either by ID or an alternative value. A
good example of this Organization which is passed in to many List()
methods.

This change enables all requests which do not have a body to lookup
object by alternative fields by supplying the lookup arguments via the
query string, rather than as part of the request path.