feat(katapult): add New<resource>Lookup() functions
For resource types which can be looked up / queried based on multiple
fields, these new functions allows creating a shallow struct which just
contains the relevant lookup field.

Most methods which accept resources as input, like the
`Create(context.Context, *Organization, *IPAddressCreateArguments)`
function on the `IPAddressesClient`, only uses the *Organization to
figure out which organization to create the IP within. This means it
does not need to be a complete *Organization with all fields populated,
but instead it can be a shallow struct containing only a `ID` or
`SubDomain` value.

The `NewOrganizationLookup()` function accepts a string that can either
be a organization ID (formatted as "org_<random>"), or a organization
sub-domain (formatted as a DNS Label). Based on the format of the given
string, the returned *Organization will have the given string populated
as either the ID or SubDomain field.