fix(load_balancer): disable https_redirect
Trying to update load balancers to set the HTTPS Redirect property to
false was not possible, as a false value would cause the field to not be
included in the marshaled JSON payload, which the API interprets
as "no change".

Hence the field type needs to be `*bool` so we can distinguish between
`true`, `false`, and nil/missing.