fix(load_balancer): enable removing all certs from a rule
Previously when setting the Certificates field to a empty slice, the
field would not be sent. This prevented making a request to remove all
certificates assigned to a rule. Hence we need to make the Certificates
field a slice pointer, so we can differentiate between not provided (no
change), and empty list (remove all).