test(testclient): improve usage of testclient in tests
Previously we were not actually testing the manipulation of the
katapult.Response object that paginated list endpoints perform. As the
pagination data comes from parsing the main response body, it must be
manually assigned to the katapult.Response object within the List()
function.

This assignment could be completely removed and the existing tests we
had in place would happily carry on working. Now the response that the
testclient returns, and the wanted/expected response are two separate
things, allowing us to validate that the List() function is actually
populating the the Pagination field on the Response object.