fix(codegen): fix issue caused by a recent change to Katapult's API Schema
The underlying project that Katapult for it's API was renamed from Rapid
to Apia (and open-sourced). This changed some internal naming
conventions in the schema file produced by Katapult's API, which in turn
broke codgen here in go-katapult, due to the basic types living within
the "Apia" top-level namespace instead of "Rapid" as they used to.

This change makes codegen compatible with both old and new schema
formats.