chore(codegen): bundle Katapult API schema into go-katapult git repo
Code generation currently automatically fetches the latest version of
Katapult's API schema, this has led to annoyances over the past couple
of months when the schema changes.

Instead we now base code generation on a snapshot of the API schema, and
run a nightly task to check if the bundled schema is up to date or not.