chore(codegen): warn about error code vs name miss-matches
For example the error "ErrVirtualMachineBuildNotFound" has a error
code of "build_not_found", while we would expect the code to be
"virtual_machine_build_not_found", or the error's ID to end with
BuildNotFound.

This should help visualize potential oddities with the Katapult Core API
schema which we might need to deal with.